Move config to auth package
