+++ cmd/auth/logout.go
@@ -2,8 +2,7 @@ package auth
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"github.com/tensorleap/cli-go/pkg/config"
+	"github.com/tensorleap/cli-go/pkg/auth"
 )
 
 func init() {
@@ -12,8 +11,7 @@ func init() {
 		Short: "Remove api key from the machine",
 		Long:  `Remove api key from the machine`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			viper.Set(config.API_KEY_CONFIG_PATH, "")
-			return viper.WriteConfig()
+			return auth.Logout()
 		},
 	})
 }
