+++ cmd/datasets/list.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 	. "github.com/tensorleap/cli-go/pkg/api"
-	"github.com/tensorleap/cli-go/pkg/config"
+	"github.com/tensorleap/cli-go/pkg/auth"
 )
 
 func init() {
@@ -14,7 +14,7 @@ func init() {
 		Short: "List available datasets",
 		Long:  `List available datasets`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := config.CheckLoggedIn(); err != nil {
+			if err := auth.CheckLoggedIn(); err != nil {
 				return err
 			}
 			data, _, err := ApiClient.GetDatasets(cmd.Context()).Execute()
