+++ cmd/datasets/listVersions.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 	. "github.com/tensorleap/cli-go/pkg/api"
-	"github.com/tensorleap/cli-go/pkg/config"
+	"github.com/tensorleap/cli-go/pkg/auth"
 	"github.com/tensorleap/cli-go/pkg/tensorleapapi"
 )
 
@@ -17,7 +17,7 @@ func init() {
 		Short: "List available dataset versions",
 		Long:  `List available dataset versions`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := config.CheckLoggedIn(); err != nil {
+			if err := auth.CheckLoggedIn(); err != nil {
 				return err
 			}
 			params := *tensorleapapi.NewGetDatasetVersionsParams(datasetId)
